working with two reducers but is kind of smelly - combineReducers not yet implemented
