+++ src/App.js
@@ -51,7 +51,7 @@ class App extends React.Component {
 }
 const mapStateToProps = state => {
   return {
-    theme: state.theme
+    theme: state.themeR.theme
   };
 }
